[#47] Pull request fixes.
