+++ deploy/ansible/roles/legion_core_chart/templates/legion-core-values.yaml.j2
@@ -35,10 +35,10 @@ jenkins:
 
     dags_volume_pvc: 'legion-airflow-shared' # from shared_k8s_resources/templates
 
-liara:
+console:
   enabled: true
   image:
-    repository: "{{ nexus_docker_repo }}/legion/k8s-liara"
+    repository: "{{ nexus_docker_repo }}/legion/k8s-console"
     tag: "{{ base_version }}-{{ local_version }}"
 
 graphite:
