+++ deploy/helms/legion-core/templates/console/console-secrets.yaml
@@ -1,0 +1,23 @@
+{{ if .Values.console.enabled -}}
+kind: ConfigMap
+apiVersion: v1
+metadata:
+  name: "{{ .Release.Name }}-cluster-state"
+data:
+  cluster.yaml: |
+    namespace: "{{ .Release.Namespace }}"
+    deployment: "{{ .Release.Name }}"
+    domain: "{{ .Values.domainDelimiter }}{{ .Values.rootDomain }}"
+    dashboard:
+      public: "dashboard{{ .Values.domainDelimiter }}{{ .Values.rootDomain }}"
+      domain: "{{ .Release.Name }}-dashboard.{{ .Release.Namespace }}.svc.{{ .Values.localDomain }}"
+      port: 80
+    grafana:
+      public: "grafana{{ .Values.domainDelimiter }}{{ .Values.rootDomain }}"
+      domain: "{{ .Release.Name }}-grafana.{{ .Release.Namespace }}.svc.{{ .Values.localDomain }}"
+      port: 80
+    jenkins:
+      public: "jenkins{{ .Values.domainDelimiter }}{{ .Values.rootDomain }}"
+      domain: "{{ .Release.Name }}-jenkins.{{ .Release.Namespace }}.svc.{{ .Values.localDomain }}"
+      port: 80
+{{- end }}
