--- deploy/helms/legion-core/templates/liara/liara-rbac.yaml
@@ -1,31 +1,0 @@
-{{ if .Values.liara.enabled -}}
-{{ if .Values.liara.rbac -}}
-kind: ClusterRole
-apiVersion: rbac.authorization.k8s.io/v1beta1
-metadata:
-  name: liara
-rules:
-- apiGroups: [""] # "" indicates the core API group
-  resources: ["namespaces", "services"]
-  verbs: ["get", "watch", "list"]
----
-kind: ClusterRoleBinding
-apiVersion: rbac.authorization.k8s.io/v1beta1
-metadata:
-  name: liara
-subjects:
-- kind: ServiceAccount
-  namespace: {{ .Release.Namespace }}
-  name: liara
-roleRef:
-  kind: ClusterRole
-  name: liara
-  apiGroup: rbac.authorization.k8s.io
----
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: liara
-{{- end }}
-{{- end }}
-
