+++ legion/bin/console
@@ -1,0 +1,54 @@
+#!/usr/bin/env python
+#
+#    Copyright 2018 EPAM Systems
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+#
+
+try:
+    import docker_bootup
+except ImportError:
+    pass
+
+import sys
+import os
+import argparse
+import logging
+
+from legion.console.server import CONSOLE
+import legion.utils
+
+ROOT_LOGGER = logging.getLogger()
+
+if __name__ == '__main__':
+    parser = argparse.ArgumentParser(description='legion Api server')
+    parser.add_argument('--verbose',
+                        help='verbose log output',
+                        action='store_true')
+
+    parser.add_argument('--legion-api-addr', type=str)
+    parser.add_argument('--legion-api-port', type=int)
+
+    # --------- END OF SECTIONS -----------
+    args = parser.parse_args(sys.argv[1:])
+
+    v = vars(args)
+
+    if args.verbose or legion.utils.string_to_bool(os.getenv('VERBOSE', '')):
+        log_level = logging.DEBUG
+    else:
+        log_level = logging.ERROR
+
+    ROOT_LOGGER.setLevel(log_level)
+
+    CONSOLE.start(args)
