+++ legion/legion/console/plugins/hello/static/css/hello.css
@@ -1,0 +1,6 @@
+#message {
+    display: inline-block;
+    margin-top: 50vh;
+    margin-left: 40%;
+    font-size: 30px
+}
