+++ legion/legion/console/plugins/homepage/static/css/homepage.css
@@ -1,0 +1,5 @@
+#dashboard {
+    width:100%;
+    height:100%;
+    border:none;
+}
