+++ legion/legion/console/templates/layout.html
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <title>Console Admin Panel</title>
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/layout.css') }}">
+  </head>
+  <body>
+    <div>
+      <div id="main_menu">
+        <h2>Console</h2>
+
+        <h4>Cluster-wide</h4>
+        <ul>
+          {% for main_item in main_menu_items %}
+            <li>{{ main_item }}</li>
+          {% endfor %}
+        </ul>
+
+        {% for enclave_name in enclaves_menu_items.keys() %}
+        <h4>{{ enclave_name }}</h4>
+          <ul>
+            {% for item in enclaves_menu_items[enclave_name] %}
+              <li>{{ item }}</li>
+            {% endfor %}
+          </ul>
+        {% endfor %}
+
+      </div>
+      <div id="content">
+        {% block content %}
+        {% endblock %}
+      </div>
+    </div>
+  </body>
+</html>
