--- legion/legion/liara/plugins/hello/static/css/hello.css
@@ -1,6 +1,0 @@
-#message {
-    display: inline-block;
-    margin-top: 50vh;
-    margin-left: 40%;
-    font-size: 30px
-}
