--- legion/legion/liara/plugins/hello/templates/hello.html
@@ -1,5 +1,0 @@
-{% extends "layout.html" %}
-{% block content %}
-<link rel="stylesheet" href="{{ url_for('hello.static', filename='css/hello.css') }}">
-<div id="message">Hello {{enclave}}!</div>
-{% endblock %}
