--- legion/legion/liara/plugins/homepage/templates/homepage.html
@@ -1,5 +1,0 @@
-{% extends "layout.html" %}
-{% block content %}
-<link rel="stylesheet" href="{{ url_for('homepage.static', filename='css/homepage.css') }}">
-<iframe id="dashboard" src="{{ dashboard_url }}"></iframe>
-{% endblock %}
